double -> float

+++ include/cpu/icp.hpp
@@ -9,21 +9,21 @@
 #include "cpu/tuple.hpp"
 
 std::vector<std::tuple<size_t, int>> get_correspondence_indices(CPUMatrix &P, CPUMatrix &Q);
-std::vector<std::tuple<size_t, int>> get_correspondence_indices(double *P, double *Q, size_t P_r,
+std::vector<std::tuple<size_t, int>> get_correspondence_indices(float *P, float *Q, size_t P_r,
                                                                  size_t P_c, size_t Q_r, size_t Q_c);
-void get_correspondence_indices_array(tuple **correspondances, double *P, double *Q, size_t P_r, size_t P_c, size_t Q_r, size_t Q_c);
+void get_correspondence_indices_array(tuple **correspondances, float *P, float *Q, size_t P_r, size_t P_c, size_t Q_r, size_t Q_c);
 
-double default_kernel(CPUMatrix a);
-double default_kernel(double a);
-void increment_cov(double *P, double *Q);
+float default_kernel(CPUMatrix a);
+float default_kernel(float a);
+void increment_cov(float *P, float *Q);
 
-std::tuple<CPUMatrix, std::vector<double>> compute_cross_variance(CPUMatrix &P, CPUMatrix &Q,
-    const std::vector<std::tuple<size_t, int>>& correspondences, double (*kernel)(CPUMatrix a));
+std::tuple<CPUMatrix, std::vector<float>> compute_cross_variance(CPUMatrix &P, CPUMatrix &Q,
+    const std::vector<std::tuple<size_t, int>>& correspondences, float (*kernel)(CPUMatrix a));
 
-double *compute_cross_variance_cpu_call_gpu(double *P, double *Q, std::vector<std::tuple<size_t, int>> correspondences, size_t P_r, size_t P_c,
+float *compute_cross_variance_cpu_call_gpu(float *P, float *Q, std::vector<std::tuple<size_t, int>> correspondences, size_t P_r, size_t P_c,
                                 size_t Q_r, size_t Q_c);
 
-void compute_cross_variance_array(double * cov, double *P, double *Q, std::tuple<size_t, int> *correspondences, size_t P_r, size_t P_c,
+void compute_cross_variance_array(float * cov, float *P, float *Q, std::tuple<size_t, int> *correspondences, size_t P_r, size_t P_c,
                                 size_t Q_r, size_t Q_c);
 
-std::tuple<CPUMatrix, std::vector<double>, std::vector<std::tuple<size_t, int>>> icp(CPUMatrix &P, CPUMatrix &Q, unsigned iterations);
+std::tuple<CPUMatrix, std::vector<float>, std::vector<std::tuple<size_t, int>>> icp(CPUMatrix &P, CPUMatrix &Q, unsigned iterations);
