+++ include/libalg/print.hpp
@@ -13,7 +13,7 @@
  ** \param n the number of lines
  ** \param lda the number of columns per line in memory, can be != than m
  **/
-void print_matrix(std::ostream &s, double *a, int m, int n, int lda);
+void print_matrix(std::ostream &s, float *a, int m, int n, int lda);
 
 /**
  ** \brief print_matrix outputs a given matrix to the given stream
@@ -25,4 +25,4 @@ void print_matrix(std::ostream &s, double *a, int m, int n, int lda);
  **
  ** \note this calls the print_matrix(... lda) function with lda = m
  **/
-void print_matrix(std::ostream &s, double *a, int m, int n);
+void print_matrix(std::ostream &s, float *a, int m, int n);
